RDKC-9515 RDKB-38648 ARRISXB3-11307 - Random failure with testapp
Reason for change: Addressed the random failures faced with RDKC and ARRISXB3.
Also changes made for varying number of test cases in each run.
Test Procedure: Test and verified
Risks: Low

+++ sampleapps/provider/rbusSampleProvider.c
@@ -374,7 +374,7 @@ rbusError_t SampleProvider_allTypesSetHandler(rbusHandle_t handle, rbusProperty_
         int bytes_len=0;
         uint8_t const* ptr = rbusValue_GetBytes(value, &bytes_len);
         memset(&gTestSampleVal.m_bytes, 0, sizeof(gTestSampleVal.m_bytes));
-        if(bytes_len > sizeof(gTestSampleVal.m_bytes))
+        if((size_t)bytes_len > sizeof(gTestSampleVal.m_bytes))
             bytes_len = sizeof(gTestSampleVal.m_bytes);
         memcpy(&gTestSampleVal.m_bytes, ptr, bytes_len);
     }
